Added CDNCache class to provide CDN cache related functionalities
- Added a new class CDNCache that provides CDN cache flush method.
  This was inherited in publish and clear-repo commands.
- Similar class was removed from task.py
- Refactored logging and comments in publish command impl
